api_google/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai>=0.3.0
google-cloud-storage>=2.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
requests==2.32.5
asyncpg==0.29.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
redis==5.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
